requirements.txt
asgiref==3.5.2
backcall==0.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.16
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-markdownx==3.0.1
idna==3.4
importlib-metadata==5.2.0
ipython==7.34.0
jedi==0.18.2
Markdown==3.4.1
matplotlib-inline==0.1.6
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
traitlets==5.7.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
zipp==3.11.0
